#48 Restructure helpers and interfaces
 * To better align with best practices
