+++ composables/useAvatar.ts
@@ -1,0 +1,3 @@
+export function useAvatar(username: string = 'CODE-CLAN-AUS'): string {
+  return `https://github.com/${username}.png`;
+}
