--- src/helpers/avatar.ts
@@ -1,3 +1,0 @@
-export function generateAvatarUrl(username: string = 'CODE-CLAN-AUS'): Promise<string> {
-  return `https://github.com/${username}.png`;
-}
