+++ types/IRatingTableRow.ts
@@ -1,0 +1,6 @@
+export interface IRatingTableRow {
+  key: number;
+  label: string;
+  percentage: number;
+  votes: number;
+}
